SRC/microservices/services/llm-hub/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
redis==5.0.1
google-generativeai==0.3.2
pydantic==2.5.0
python-multipart==0.0.6